Add plugin:// path to stitch together images
