+++ addon.xml
@@ -15,6 +15,8 @@
   <extension point="xbmc.python.script"         library="default.py">
     <provides>executable</provides>
   </extension>
+  <!-- Kodi seems to pass xbmc.python.[script|library] to the same script even if defined otherwise, but it's easy to detect in the script -->
+  <extension point="xbmc.python.library" library="default.py" />
   <extension point="xbmc.addon.metadata">
     <summary lang="be">Downloads Artwork for TV shows, Movies and Musicvideos in your library</summary>
     <summary lang="da">Downloader grafik til TV-serier, Film og Musikvideoer i dit bibliotek</summary>
