+++ lib/provider/fanarttv.py
@@ -23,13 +23,14 @@ import urllib
 
 ### import libraries
 #from lib.provider.base import BaseProvider
+from lib import common
 from lib.language import *
 from lib.script_exceptions import NoFanartError
 from lib.utils import *
 from operator import itemgetter
 
 ### get addon info
-__localize__    = ( sys.modules[ '__main__' ].__localize__ )
+__localize__ = common.__localize__
 
 API_KEY = '586118be1ac673f74963cc284d46bd8e'
 API_URL_TV = 'http://webservice.fanart.tv/v3/tv/%s?api_key=%s'
